requirements.txt
# LLM4AscendC：eval_operator、MKB reference、pybind 扩展构建与 NPU 评测
# Python 建议：3.10（与下列 wheel 标签 cp310 一致）
# 版本快照来源：Ascend 开发容器内 pip 列表（torch 2.1.0 + CANN 配套 torch_npu）
#
# torch_npu 不在 PyPI 公开发布，需使用与 CANN / 驱动匹配的官方 wheel，例如：
#   pip install /path/to/torch_npu-2.1.0.post12-cp310-cp310-linux_aarch64.whl
# 具体文件名以本机 ascend-toolkit 随附包为准。

# --- PyPI 可固定版本 ---
ninja==1.11.1.4
numpy==1.26.0
packaging==25.0
pybind11==2.13.6
setuptools==65.5.0
torch==2.1.0
typing_extensions==4.15.0
wheel==0.45.1

# --- LangGraph Agent（generator/agent 智能体模块）---
langgraph
langchain-openai
langchain-core
openai

# --- Web 搜索（可选，用于 agent web 检索）---
ddgs
trafilatura
requests

# --- KB 知识库（可选，用于 ChromaDB + LlamaIndex + BGE-M3）---
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb

# --- RAG 代码检索 ---
sentence-transformers

# --- PDF 解析（KB 构建，可选）---
unstructured
pdfminer.six
pypdf
